# services/api-gateway/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the CareerUP API server.",
        "title": "CareerUP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/ilo/result": {
            "post": {
                "description": "Submit ILO test result for the authenticated user and get analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ilo"
                ],
                "summary": "Submit ILO test result",
                "parameters": [
                    {
                        "description": "ILO Test Result Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.IloTestResultRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.IloTestResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ilo/result/{id}": {
            "get": {
                "description": "Get a specific ILO test result by ID for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ilo"
                ],
                "summary": "Get a specific ILO test result by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Result ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.IloTestResultResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ilo/results": {
            "get": {
                "description": "Get all ILO test results for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ilo"
                ],
                "summary": "Get all ILO test results for a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.IloTestResultResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "client.IloDomainScore": {
            "type": "object",
            "properties": {
                "domain_code": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "percent": {
                    "type": "number"
                },
                "rank": {
                    "type": "integer"
                },
                "raw_score": {
                    "type": "integer"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "handler.IloAnswer": {
            "type": "object",
            "properties": {
                "question_id": {
                    "type": "string"
                },
                "question_number": {
                    "type": "integer"
                },
                "selected_option": {
                    "type": "integer"
                }
            }
        },
...
        "handler.IloTestResultRequest": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.IloAnswer"
                    }
                },
                "result_data": {
                    "type": "string",
                    "example": "{\"score\":85,\"details\":{...}}"
                }
            }
        },
        "handler.IloTestResultResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "result_data": {
                    "type": "string"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/client.IloDomainScore"
                    }
                },
                "suggested_careers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "top_domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "in": "header",
          "name": "Authorization"
        }
    },
    "security": [
        {
          "BearerAuth": []
        }
    ]
}